refactor: split armor layout engine helpers
